+++ renku/command/update.py
@@ -32,8 +32,8 @@ from renku.core.workflow.activity import (
     is_activity_valid,
     sort_activities,
 )
-from renku.core.workflow.concrete_execution_graph import ExecutionGraph
 from renku.core.workflow.execute import execute_workflow_graph
+from renku.core.workflow.model.concrete_execution_graph import ExecutionGraph
 from renku.domain_model.project_context import project_context
